[OC] fix global alt.scaling

+++ software/o_c_REV/OC_DAC.h
@@ -59,7 +59,7 @@ public:
   static void reset_all_auto_channel_calibration_data();
   static void choose_calibration_data();
   static void set_scaling(uint8_t scaling, uint8_t channel_id);
-  static void restore_scaling(uint8_t scaling);
+  static void restore_scaling(uint32_t scaling);
   static uint8_t get_voltage_scaling(uint8_t channel_id);
   static uint32_t store_scaling();
